sdy/rag-llamaindex/pyproject.toml
[project]
name = "rag-llamaindex"
version = "0.1.0"
description = "RAG Pipeline with LlamaIndex framework"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-vector-stores-chroma>=0.3.0",
    "llama-index-readers-file>=0.3.0",
    "chromadb>=0.4.0",
    "typer>=0.9.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
rag-li = "rag_llamaindex.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_llamaindex"]

# --- Ruff ---
[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"

# --- Pytest ---
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
